+++ fltk-sys/src/widget.rs
@@ -172,6 +172,22 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Widget_resize_callback(
+        self_: *mut Fl_Widget,
+        cb: ::std::option::Option<
+            unsafe extern "C" fn(
+                arg1: *mut Fl_Widget,
+                x: ::std::os::raw::c_int,
+                y: ::std::os::raw::c_int,
+                w: ::std::os::raw::c_int,
+                h: ::std::os::raw::c_int,
+                arg2: *mut ::std::os::raw::c_void,
+            ),
+        >,
+        data: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
     pub fn Fl_Widget_set_when(arg1: *mut Fl_Widget, arg2: ::std::os::raw::c_int);
 }
 extern "C" {