+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## [1.2.0] - Unreleased
 - Add FL_FLex as submodule and expose a group::Flex widget.
+- Add WidgetBase::resize_callback().
 
 ## [1.1.19] - 2021-09-18
 - Fix doc comments about FileDialog and NativeFileChooser.